Refactoring Rando WIP act 2

+++ sadx-randomizer/data/chao.h
@@ -16,7 +16,7 @@ enum ChaoActions {
 
 struct LostChaoPosition
 {
-	short version;
+	int8_t version;
 	int16_t LevelID;
 	NJS_VECTOR Position;
 	float YRot;
